Resolved template errors
I changed the creation of new nodes as it was irrelevant to pass a node to the list in the main function, instead I generate a new node inside the Append function and use data as the only parameter. Resolved missusage of an Add Head and Add Node function, they were both the same generation process, so I just merged them with a conditional that checks if head exists, and as I am no longer passing a node I renamed the whole function to append.